test(cli): refactor import path in extentsion tests
